+++ App.js
@@ -17,6 +17,10 @@ import ChatInfo from "./src/screens/Home/ChatInfo";
 import ChatInfoGroup from "./src/screens/Home/ChatInfoGroup";
 import AddMembers from "./src/screens/Home/AddMembers";
 import SeeMembers from "./src/screens/Home/SeeMembers";
+import EditBubble from "./src/screens/Home/EditBubble";
+import EditMembers from "./src/screens/Home/EditMembers";
+import ExcludeMembers from "./src/screens/Home/ExcludeMembers";
+import CreateAudience from "./src/screens/Home/CreateAudience";
 
 const Stack = createStackNavigator();
 const AuthenticatedUserContext = createContext({});
@@ -48,6 +52,10 @@ const ChatStack = () => {
       <Stack.Screen name="ChatInfoGroup" component={ChatInfoGroup} />
       <Stack.Screen name="AddMembers" component={AddMembers} />
       <Stack.Screen name="SeeMembers" component={SeeMembers} />
+      <Stack.Screen name="EditBubble" component={EditBubble} />
+      <Stack.Screen name="EditMembers" component={EditMembers} />
+      <Stack.Screen name="ExcludeMembers" component={ExcludeMembers} />
+      <Stack.Screen name="CreateAudience" component={CreateAudience} />
     </Stack.Navigator>
   );
 };