+++ src/components/AudienceBox.js
@@ -3,7 +3,7 @@ import { Image, View, Text, TouchableOpacity, TouchableWithoutFeedback } from "r
 
 import useIcon from "../hooks/useIcon";
 import audienceBoxStyles from "./utils/audienceBoxStyles";
-
+import DeleteAlertPopup from "./DeleteAlertPopup";
 
 const AudienceBox = (props) => {
     const editIcon = useIcon("editIcon");
@@ -19,8 +19,12 @@ const AudienceBox = (props) => {
 
     const openCloseIcon = isExcludePersonVisible ? openIcon : closeIcon;
 
+
+    const [isDeleteAlertPopupVisible, setDeleteAlertPopupVisible] = useState(false);
+
     return (
         <View style={audienceBoxStyles.audienceBoxContainer}>
+            <DeleteAlertPopup is2ndModalVisible={isDeleteAlertPopupVisible} set2ndModalVisible={setDeleteAlertPopupVisible} set1stModalVisible={setDeleteAlertPopupVisible} deleteWhatText={"delete this group"}/>
             <View>
                 <Image 
                     source={props.audienceIndicator}
@@ -85,14 +89,14 @@ const AudienceBox = (props) => {
                 <Text style={audienceBoxStyles.audienceBoxClearStatusText}>Status will clear after tomorrow 8:30 PM</Text>
             </View>
             <View style={audienceBoxStyles.audienceBoxIconsContainer}>
-                <TouchableOpacity>
+                <TouchableOpacity onPress={() => props.navigation.navigate("EditBubble")}>
                     <Image
                         source={editIcon}
                         style={audienceBoxStyles.audienceBoxEditIcon}
                     />
                 </TouchableOpacity>
 
-                <TouchableOpacity>
+                <TouchableOpacity onPress={() => setDeleteAlertPopupVisible(true)}>
                     <Image 
                         source={trashIcon}
                         style={audienceBoxStyles.audienceBoxTrashIcon}